[WIP] Pixel perfect mouse position.

+++ Graphics Engine/Game/Include/States/InputState.h
@@ -11,7 +11,7 @@ namespace game
 			released
 		};
 
-		glm::vec2 mousePos{};
+		glm::ivec2 mousePos{};
 		State lMouse{};
 		State rMouse{};
 		float scroll = 0;
